+++ packages/architectura/src/Core/Server.mts
@@ -19,11 +19,12 @@ import { ExecutionContextService } from "../service/execution-context/execution-
 
 
 
+import { KernelService } from "../service/kernel/kernel.service.mjs";
+
 import { RichClientRequest } from "./RichClientRequest.mjs";
 
 import { RichServerResponse } from "./RichServerResponse.mjs";
 
-import { KernelService } from "./kernel.service.mjs";
