+++ packages/architectura/src/Core/index.mts
@@ -1,10 +1,10 @@
 export * from "../service/execution-context/execution-context.service.mjs";
 export * from "../service/executor/executor.service.mjs";
+export * from "../service/kernel/kernel.service.mjs";
 export * from "./Base/index.mjs";
 export * from "./ExecutionContext/index.mjs";
 export * from "./Executor/index.mjs";
 export * from "./HTTP/index.mjs";
-export * from "./kernel.service.mjs";
 export * from "./RichClientRequest.mjs";
 export * from "./RichServerResponse.mjs";
 export * from "./Server.mjs";
