+++ packages/architectura/src/service/_index.mts
@@ -1,7 +1,9 @@
 export * from "./cookie/_index.mjs";
 export * from "./dispatcher/_index.mjs";
 export * from './execution-context/_index.mjs';
+export * from './executor/_index.mjs';
 export * from "./file-system/_index.mjs";
+export * from "./kernel/_index.mjs";
 export * from "./logger/_index.mjs";
 export * from "./session-manager/_index.mjs";
 export * from "./session/_index.mjs";
