+++ packages/architectura/src/service/kernel/_index.mts
@@ -1,0 +1,1 @@
+export * from './kernel.service.mjs';
